+++ js/form.js
@@ -1,17 +1,11 @@
 import {sendData} from './data.js';
-
 import {
   showSuccessMessage,
   showErrorMessage
 } from './popup.js';
-
 import {map, mainPinMarker, TokyoCoordinate} from './map.js';
 
 const adForm = document.querySelector('.ad-form');
-const adFormItems = adForm.querySelectorAll('fieldset');
-const mapFilters = document.querySelector('.map__filters');
-const mapFiltersItems = mapFilters.children;
-
 const titleForm = adForm.querySelector('#title');
 const priceForm = adForm.querySelector('#price');
 const roomForm = adForm.querySelector('#room_number');
@@ -29,34 +23,6 @@ const pristine = new Pristine(adForm, {
   errorTextParent: 'ad-form__element',
   errorTextClass: 'text-help',
 });
-
-const disableAdForm = () => {
-  adForm.classList.add('ad-form--disabled');
-  for (const element of adFormItems) {
-    element.setAttribute('disabled', 'disabled');
-  }
-};
-
-const disableMapFilters = () => {
-  mapFilters.classList.add('map__filters--disabled');
-  for (const element of mapFiltersItems) {
-    element.setAttribute('disabled', 'disabled');
-  }
-};
-
-const enableAdForm = () => {
-  adForm.classList.remove('ad-form--disabled');
-  for (const element of adFormItems) {
-    element.removeAttribute('disabled');
-  }
-};
-
-const enableMapFilters = () => {
-  mapFilters.classList.remove('map__filters--disabled');
-  for (const element of mapFiltersItems) {
-    element.removeAttribute('disabled');
-  }
-};
 
 // Валидация заголовка
 const validateTitle = (value) => value.length >= 30 && value.length <= 100;
@@ -114,6 +80,10 @@ timeoutForm.addEventListener('change', () => {
 });
 
 // Reset
+const setCoordinates = (coordinates) => {
+  addressForm.value = `${coordinates.lat.toFixed(5)}, ${coordinates.lng.toFixed(5)}`;
+};
+
 const resetForm = () => {
   adForm.reset();
   mainPinMarker.setLatLng({
@@ -126,7 +96,9 @@ const resetForm = () => {
   }, 10);
   map.closePopup();
   slider.noUiSlider.reset();
+  setCoordinates(mainPinMarker.getLatLng());
 };
+
 
 const onResetButtonClick = (evt) => {
   evt.preventDefault();
@@ -155,13 +127,9 @@ const onFormSubmit = (evt) => {
 adForm.addEventListener('submit', onFormSubmit);
 
 export {
-  disableAdForm,
-  disableMapFilters,
-  enableAdForm,
-  enableMapFilters,
   addressForm,
   priceForm,
   slider,
   typeForm,
-  TypePriceMap
+  TypePriceMap,
 };
