+++ js/main.js
@@ -1,10 +1,14 @@
 import './card.js';
 import './slider.js';
-import {
-  disableAdForm,
-  disableMapFilters
-} from './form.js';
-import './map.js';
+import {addressForm} from './form.js';
+import {enableAdForm} from './change-activity.js';
+import {map, TokyoCoordinate} from './map.js';
 
-disableAdForm();
-disableMapFilters();
+map.on('load', () => {
+  enableAdForm();
+  addressForm.value = `${TokyoCoordinate.LAT}, ${TokyoCoordinate.LNG}`;
+})
+  .setView({
+    lat: TokyoCoordinate.LAT,
+    lng: TokyoCoordinate.LNG
+  }, 10);
